Fixing bug - TIFF does not support saving bool

+++ fullconvnets/prepare.py
@@ -120,7 +120,7 @@ def crop_training_chunks():
                              step=const.STEP_3D,
                              folder=const.FOLDER_TRAIN_IMAGE_CROP_3D)
 
-    utils.save_cropped_chunk(data_label,
+    utils.save_cropped_chunk(np.uint8(data_label),
                              window_shape=const.WINDOW_SHAPE_3D,
                              step=const.STEP_3D,
                              folder=const.FOLDER_TRAIN_LABEL_CROP_3D)
@@ -146,7 +146,7 @@ def crop_training_chunks():
                              step=const.STEP_3D,
                              folder=const.FOLDER_VAL_IMAGE_CROP_3D)
 
-    utils.save_cropped_chunk(data_label,
+    utils.save_cropped_chunk(np.uint8(data_label),
                              window_shape=const.WINDOW_SHAPE_3D,
                              step=const.STEP_3D,
                              folder=const.FOLDER_VAL_LABEL_CROP_3D)
